Moved decorator to test folder
